# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/cart": {
      "get": {
        "description": "Always succeeds for an authenticated caller. A user who has never added anything gets an empty cart with both totals at `0` rather than a 404 — the cart is created lazily on the first add, and its absence is not an error.\n\nCatalogue data on each line is read **live**, not frozen: price, status and the size’s own stock are current as of this request. That is what lets a storefront warn \"only 2 left\" or \"no longer available\". Prices freeze at checkout and not before.\n\nEach line names a **variant** — `variantId` is what PATCH and DELETE address, and the stock that matters is `variant.stockQuantity`, the count for that size, not a product-wide total.\n\n`itemsSubtotalCents` and `itemCount` are computed here from those same live prices, so no client has to sum money. There is deliberately **no order total** on this route: without a postal code there is no freight, and a \"total\" that is missing the freight is precisely the number a checkout must never show. That one comes from POST /shipping/quote, as `orderTotalCents` per option.",
        "operationId": "CartController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponse"
                }
              }
            }
          },
          "401": {
            "description": "No bearer token, or the token is expired or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the current cart",
        "tags": [
          "cart"
        ]
      },
      "delete": {
        "description": "Removes every line. Idempotent — emptying an already-empty cart is a 200 with an empty list, not an error.",
        "operationId": "CartController_clear",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponse"
                }
              }
            }
          },
          "401": {
            "description": "No bearer token, or the token is expired or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Empty the cart",
        "tags": [
          "cart"
        ]
      }
    },
    "/cart/items": {
      "post": {
        "description": "Takes a **variantId** — a size — not a product id. A product id does not say which size, and this API refuses to pick one on your behalf (docs/specs/product-variants.md).\n\nAdditive: adding a variant already in the cart **increases** its quantity rather than creating a second line, while a different size of the same product is its own line. Use PATCH below to set an absolute quantity.\n\nOnly variants of ACTIVE products can be added. A variant of a DRAFT or ARCHIVED product and a variant that does not exist are the same 404 — the public does not get to tell an unreleased product from one that never existed.",
        "operationId": "CartController_addItem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCartItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The whole cart, not just the line that changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponse"
                }
              }
            }
          },
          "400": {
            "description": "`variantId` is not a UUID, or `quantity` is outside 1–999.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No bearer token, or the token is expired or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such variant, or its product is not ACTIVE.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add to the cart",
        "tags": [
          "cart"
        ]
      }
    },
    "/cart/items/{variantId}": {
      "patch": {
        "description": "Absolute — \"make it 5\" — as opposed to POST /cart/items, which adds. The line is addressed by its **variantId**. To remove a line use DELETE; zero is not a valid quantity here.",
        "operationId": "CartController_setQuantity",
        "parameters": [
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetCartItemQuantityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponse"
                }
              }
            }
          },
          "400": {
            "description": "`quantity` is outside 1–999.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No bearer token, or the token is expired or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "That variant is not in the cart.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set a line to an absolute quantity",
        "tags": [
          "cart"
        ]
      },
      "delete": {
        "operationId": "CartController_removeItem",
        "parameters": [
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponse"
                }
              }
            }
          },
          "401": {
            "description": "No bearer token, or the token is expired or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "That variant is not in the cart.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove one line from the cart",
        "tags": [
          "cart"
        ]
      }
    },
...
  },
  "info": {
    "title": "commerce-core",
...
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "The `accessToken` from /auth/login, /auth/refresh or the Google callback."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 409,
            "description": "Mirrors the HTTP status code."
          },
          "message": {
            "description": "A human-readable reason, or one entry per failed validation rule.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "examples": [
              "Order is PAID; only a CREATED order can be paid",
              [
                "priceCents must be an integer number",
                "name should not be empty"
              ]
            ]
          },
          "error": {
            "type": "string",
            "example": "Conflict",
            "description": "The status code's canonical name."
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "CartProductResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Camiseta Azul"
          },
          "slug": {
            "type": "string",
            "example": "camiseta-azul"
          },
          "priceCents": {
            "type": "number",
            "description": "Current catalog price, in cents.",
            "example": 7990
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "ARCHIVED"
            ],
            "description": "Anything other than ACTIVE means checkout will refuse this line with a 409 until it is removed."
          },
          "weightGrams": {
            "type": "number",
            "nullable": true,
            "example": 180
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "priceCents",
          "status",
          "weightGrams"
        ]
      },
      "CartVariantResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "label": {
            "type": "string",
            "example": "M"
          },
          "position": {
            "type": "number",
            "description": "The size’s display order on its product.",
            "example": 1
          },
          "stockQuantity": {
            "type": "number",
            "description": "Current stock of THIS size. Lower than the line quantity is not an error here — it becomes one at checkout, with a 409 naming the piece.",
            "example": 3
          }
        },
        "required": [
          "id",
          "label",
          "position",
          "stockQuantity"
        ]
      },
      "CartItemResponse": {
        "type": "object",
        "properties": {
          "variantId": {
            "type": "string",
            "format": "uuid",
            "description": "The line’s identity. PATCH /cart/items/{variantId} and DELETE /cart/items/{variantId} address this, not the product — two sizes of one shirt are two lines."
          },
          "quantity": {
            "type": "number",
            "minimum": 1,
            "maximum": 999,
            "example": 2
          },
          "product": {
            "$ref": "#/components/schemas/CartProductResponse"
          },
          "variant": {
            "$ref": "#/components/schemas/CartVariantResponse"
          }
        },
        "required": [
          "variantId",
          "quantity",
          "product",
          "variant"
        ]
      },
      "CartResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItemResponse"
            }
          },
          "itemsSubtotalCents": {
            "type": "number",
            "description": "Sum of `product.priceCents × quantity` over the lines, in cents, computed server-side against LIVE catalogue prices. Reprice a product and this follows on the next read — the cart freezes nothing; checkout does. An empty cart is `0`, never null and never absent.",
            "example": 12480
          },
          "itemCount": {
            "type": "number",
            "description": "Sum of quantities — pieces, not lines. This is the cart badge: two shirts and a pair of trousers is 3, not 2. An empty cart is `0`.",
            "example": 3
          }
        },
        "required": [
          "items",
          "itemsSubtotalCents",
          "itemCount"
        ]
      },
      "AddCartItemDto": {
        "type": "object",
        "properties": {
          "variantId": {
            "type": "string",
            "format": "uuid",
            "description": "The **variant** (size) to add — from `ProductResponse.variants[].id`. Its product must exist and be ACTIVE, or the answer is 404. A product id is not accepted: it does not say which size."
          },
          "quantity": {
            "type": "number",
            "minimum": 1,
            "maximum": 999,
            "description": "How many to ADD. The same variant already in the cart has its quantity increased rather than duplicated; a DIFFERENT size of the same product is a separate line. The 999 bound applies per request, not to the resulting line.",
            "example": 2
          }
        },
        "required": [
          "variantId",
          "quantity"
        ]
      },
      "SetCartItemQuantityDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number",
            "minimum": 1,
            "maximum": 999,
            "description": "The absolute quantity for this line. Zero is not valid — use DELETE to remove the line.",
            "example": 5
          }
        },
        "required": [
          "quantity"
        ]
      },
...
    }
  }
}